creates library
library will hold functions for loom
